optimize connection and split cli file into mutiple files
